+++ src/VirtualAMHS/VirtualOHT.h
@@ -7,8 +7,8 @@ public:
 	virtual ~VirtualOHT(void);
 	
 public:
-	int Auth( int nPos, int nHand);
-	int UpdatePos(uint16 nPos);
+	int Auth( uint32 nPos, int nHand);
+	int UpdatePos(uint32 nPos);
 	virtual void HandleCommand(AMHSPacket& packet);
 
 private:
@@ -22,7 +22,7 @@ private:
 
 public:
 	int m_nHand;
-	int m_nPos;
+	DWORD m_nPos;
 	int m_nPosUpdateTime;
 	int m_nStatusUpdateTime;
 	int m_nPosTimerID;
